pyproject.toml
[tool.poetry]
name = "queueapi"
version = "0.1.0"
description = ""
authors = ["triks <triks0001000@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.106.0"
uvicorn = "^0.25.0"
sqlalchemy = "^2.0.23"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
django-environ = "^0.11.2"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
asyncpg = "^0.29.0"
fastapi-users = {extras = ["sqlalchemy"], version = "^12.1.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
